test_project/requirements.txt
djangoreactframework
Pillow
